+++ R/cached_data.R
@@ -159,3 +159,12 @@
 #'
 #' @usage data(cadvs)
 "cadvs"
+
+#' Cached ADAB
+#'
+#' Cached ADAB data generated with \code{seed = 1}
+#'
+#' @docType data
+#'
+#' @usage data(cadab)
+"cadab"